Fallback to OpenShift user API if no role granted
Using delegating authentication and authorization requires a cluster
role `system:auth-delegator` to be assigned to the service account the
oauth_proxy is running under. This might be difficult to add if the
cluster is not under control of the user deploying oauth_proxy.

This adds a fallback to use OpenShift user API for authentication and
ignore authorization entirely.